add tokenizer and parser
I'm not sure why, but git hub refused to add the folders containing
these files.  So I had to insert each one manually.

It might have to do with the fact that the containing folders had a
local git repository and it did some weird stuff that I wasn't prepared
for.