dh-16 The largest sum
Summing up one hundred 50 digit numbers

and they want the first 10.  You can get away with just adding
up the first 11.  By first I mean  the Most significant. which is the left most

Fact: this is basically an approximation so there may be a round error.  it is accurate
up plus or minus one.